Implemented grid management system.

+++ include/objects/components/GridComponent.h
@@ -13,14 +13,14 @@ class GridComponent : public Component {
 		const sf::Vector2i& getGridSize();
 		void setGridSize(const sf::Vector2i&);
 
-		const sf::Vector2i& getTileSize();
-		void setTileSize(const sf::Vector2i&);
+		const sf::Vector2f& getTileSize();
+		void setTileSize(const sf::Vector2f&);
 
 		GridComponent() {};
 		~GridComponent() {};
 	private:
 		sf::Vector2i gridSize = sf::Vector2i(0, 0);
-		sf::Vector2i tileSize = sf::Vector2i(0, 0);
+		sf::Vector2f tileSize = sf::Vector2f(0, 0);
 };
 
 }
